:recycle: refactor: move more interface definitions to consumers
