+++ src/__tests__/integration/server.test.ts
@@ -9,10 +9,10 @@ import { UserRepository } from "@/repositories/users/index.js";
 import { feideClient } from "@/services/auth/clients.js";
 import { FeideProvider } from "@/services/auth/providers.js";
 import { AuthService } from "@/services/auth/service.js";
-import { CabinService } from "@/services/cabins/index.js";
+import { CabinService } from "@/services/cabins/service.js";
 import { MailService } from "@/services/mail/index.js";
 import { OrganizationService } from "@/services/organizations/service.js";
-import { UserService } from "@/services/users/index.js";
+import { UserService } from "@/services/users/service.js";
 
 import { initServer } from "../../server.js";
