+++ src/graphql/auth/__tests__/integration/auth.test.ts
@@ -17,10 +17,10 @@ import { initServer } from "@/server.js";
 import { AuthClient, UserInfo } from "@/services/auth/clients.js";
 import { FeideProvider } from "@/services/auth/providers.js";
 import { AuthService } from "@/services/auth/service.js";
-import { CabinService } from "@/services/cabins/index.js";
+import { CabinService } from "@/services/cabins/service.js";
 import { MailService } from "@/services/mail/index.js";
 import { OrganizationService } from "@/services/organizations/service.js";
-import { UserService } from "@/services/users/index.js";
+import { UserService } from "@/services/users/service.js";
 
 class MockFeideClient implements AuthClient {
   fetchUserInfo(): Promise<UserInfo> {
