--- src/repositories/index.ts
@@ -1,3 +1,0 @@
-import { ICabinRepository, IUserRepository } from "./interfaces.js";
-
-export { ICabinRepository, IUserRepository };
