--- src/repositories/interfaces.ts
@@ -1,4 +1,0 @@
-import { ICabinRepository } from "./cabins/interfaces.js";
-import { IUserRepository } from "./users/interfaces.js";
-
-export { ICabinRepository, IUserRepository };
