+++ src/server.ts
@@ -21,9 +21,9 @@ import { envToLogger } from "./lib/fastify.js";
 import { migrationHealthCheck } from "./lib/prisma.js";
 import { fastifyApolloSentryPlugin } from "./lib/sentry.js";
 import { AuthService } from "./services/auth/service.js";
-import { CabinService } from "./services/cabins/index.js";
+import { CabinService } from "./services/cabins/service.js";
 import { OrganizationService } from "./services/organizations/service.js";
-import { UserService } from "./services/users/index.js";
+import { UserService } from "./services/users/service.js";
 
 interface Dependencies {
   cabinService: CabinService;
