+++ src/services/users/__tests__/unit/user.test.ts
@@ -1,28 +1,33 @@
 import { jest } from "@jest/globals";
-import { User } from "@prisma/client";
+import { Prisma, User } from "@prisma/client";
 import dayjs from "dayjs";
 import { DeepMockProxy, mockDeep } from "jest-mock-extended";
 
-import { IUserRepository } from "@/repositories/index.js";
-import { UserService } from "@/services/users/index.js";
-
-import { TestCase } from "./interfaces.js";
+import { UserRepository, UserService } from "../../service.js";
 
 const dummyUser = mockDeep<User>();
 
 const time = new Date(`${dayjs().year() + 1}-01-01`);
 
 let service: UserService;
-let repo: DeepMockProxy<IUserRepository>;
+let repo: DeepMockProxy<UserRepository>;
 
 beforeAll(() => {
-  repo = mockDeep<IUserRepository>();
+  repo = mockDeep<UserRepository>();
   service = new UserService(repo);
 
   jest.useFakeTimers().setSystemTime(time);
 });
 
 describe("UserService", () => {
+  interface TestCase {
+    name: string;
+    input: Partial<Prisma.UserUpdateInput>;
+    updateInput: Partial<Prisma.UserUpdateInput>;
+    existing: User;
+    expected: User;
+  }
+
   const testCases: TestCase[] = [
     {
       name: "set firstLogin to false",
